# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pelayanan Publik API",
        "description": "API Backend untuk Aplikasi Pelayanan Publik",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/notifikasi": {
            "get": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Get user notifications",
                "operationId": "3941d2dc81b131274651bea04ed85340",
                "parameters": [
                    {
                        "name": "jenis",
                        "in": "query",
                        "description": "Filter by jenis",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "info",
                                "peringatan",
                                "sukses",
                                "error",
                                "pengumuman"
                            ]
                        }
                    },
                    {
                        "name": "kategori",
                        "in": "query",
                        "description": "Filter by kategori",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "system",
                                "layanan",
                                "berita",
                                "saran",
                                "custom"
                            ]
                        }
                    },
                    {
                        "name": "unread_only",
                        "in": "query",
                        "description": "Show unread only",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user notifications"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifikasi/{id}": {
            "get": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Get single notification",
                "operationId": "5834fcf379923b1d1008a789244223da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Delete notification",
                "operationId": "828c453b6ffd56fe0a11c1f6e8eebd9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifikasi/{id}/read": {
            "put": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Mark notification as read",
                "operationId": "f6b4be1700affe288adb60ce1d5b871b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifikasi/read-all": {
            "put": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Mark all notifications as read",
                "operationId": "e150d1c58febd0ab0b12fa0d7d6aab73",
                "responses": {
                    "200": {
                        "description": "All notifications marked as read"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifikasi-count": {
            "get": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Get unread notification count",
                "operationId": "d2cae948908a0dff1094236e9b451fd5",
                "responses": {
                    "200": {
                        "description": "Unread notification count"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/admin/notifikasi/broadcast": {
            "post": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Send broadcast notification",
                "operationId": "7c6757c0aa1eaf8226089b9f0bf56168",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "isi"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string"
                                    },
                                    "isi": {
                                        "type": "string"
                                    },
                                    "jenis": {
                                        "type": "string",
                                        "enum": [
                                            "info",
                                            "peringatan",
                                            "sukses",
                                            "error",
                                            "pengumuman"
                                        ]
                                    },
                                    "kategori": {
                                        "type": "string",
                                        "enum": [
                                            "system",
                                            "layanan",
                                            "berita",
                                            "saran",
                                            "custom"
                                        ]
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "normal",
                                            "high",
                                            "urgent"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Broadcast notification sent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/notifikasi/{id}": {
            "delete": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Delete notification (admin)",
                "operationId": "091e3427a268903e8a311f465433a18f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/notifikasi/statistics": {
            "get": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Get notification statistics",
                "operationId": "3b7679cf65b8b5bf14747bb0b0049862",
                "responses": {
                    "200": {
                        "description": "Notification statistics"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}